--- src/components/Dashboard/Navbar.jsx
@@ -1,48 +1,0 @@
-import React,{useState} from 'react'
-import { signOut } from "firebase/auth";
-import Logo from '../../assets/logo.webp'
-import './Dashboard.css'
-
-// navbar
-function Navbar() {
-    const [menuOpen, setMenuOpen] = useState(false);
-
-  const toggleMenu = useCallback(() => {
-    setMenuOpen(!menuOpen);
-  }, [menuOpen]);
-
-  const handleSignOut = () => {
-    signOut(auth)
-      .then(() => {
-        navigate("/");
-      })
-      .catch((err) => {
-        alert(err.message);
-      });
-  };
-  return (
-    <nav className="navbar">
-          <div className="logo">
-            <img src={Logo} alt="Logo" />
-          </div>
-          <div className={`menu ${menuOpen ? 'show' : ''}`}>
-            <ul>
-              <li><a href="#">Top Universities</a></li>
-              <li><a href="#">Jobs</a></li>
-              <li><a href="#">Courses</a></li>
-              <li><a href="#">Carrier Support</a></li>
-              <li className='dot'><a href="#">•</a></li>
-              <li><a href="#" onClick={handleSignOut}>Log Out</a></li>
-              <li><a href="#"><button className='profile_btn'>Profile</button></a></li>
-            </ul>
-          </div>
-          <div className="hamburger" onClick={toggleMenu}>
-            <div className={`bar ${menuOpen ? 'open' : ''}`}></div>
-            <div className={`bar ${menuOpen ? 'open' : ''}`}></div>
-            <div className={`bar ${menuOpen ? 'open' : ''}`}></div>
-          </div>
-        </nav>
-  )
-}
-
-export default Navbar
